refactor: Move iDataQueue functions in Toolkit back to iDataQueue Class
