+++ lib/Deprecated.js
@@ -25,6 +25,7 @@
 /* eslint-disable max-classes-per-file */
 
 const deprecate = require('depd');
+const { parseString } = require('xml2js');
 
 const iPgmDeprecate = deprecate('iPgm');
 const iSqlDeprecate = deprecate('iSql');
@@ -452,24 +453,186 @@ class iConn {
 }
 
 class iDataQueue {
-  constructor(conn) {
-    iDataQueueDeprecate('As of v1.0, class \'iDataQueue\' is deprecated. Please use \'Toolkit\' instead.');
-    this.toolkit = new Toolkit(conn);
+  /**
+   * @constructor
+   * @param {Connection} conn
+   */
+  constructor(conn = {}) {
+    iDataQueueDeprecate('As of v1.0, class \'iDataQueue\' is deprecated and will be removed at a later time.');
+    if (!conn || !typeof conn === 'object') {
+      throw new Error('Expected a Connection object');
+    }
+
+    if (conn.constructor.name === 'Connection') {
+      this.conn = conn;
+      this.reportError = conn.returnError;
+    } else if (conn.constructor.name === 'iConn') {
+      this.conn = conn.connection;
+      this.reportError = this.conn.returnError;
+    } else {
+      throw new Error('Expected a Connection object');
+    }
+
+    /* when returnError is false connection.run((xmloutput))
+     * we create a new Connection object so connection.run((error, xmloutput))
+     * this.reportError is checked within the callback to notify the user or not
+    */
+    if (!this.reportError) {
+      this.conn = new Connection({
+        transport: this.conn.transport,
+        transportOptions: this.conn.transportOptions,
+      });
+    }
   }
 
+  /**
+   * @param {string} name
+   * @param {string} lib
+   * @param {string} data
+   * @param {function} cb
+   */
   sendToDataQueue(name, lib, data, cb) {
-    iDataQueueDeprecate('As of v1.0, \'iDataQueue.sendToDataQueue()\' is deprecated. Please use \'Toolkit.sendToDataQueue()\' instead.');
-    this.toolkit.sendToDataQueue(name, lib, data, cb);
+    iDataQueueDeprecate('As of v1.0, \'iDataQueue.sendToDataQueue()\' is deprecated and will be removed at a later time.');
+    const pgm = new ProgramCall('QSNDDTAQ', { lib: 'QSYS' });
+    pgm.addParam({ type: '10A', value: name });
+    pgm.addParam({ type: '10A', value: lib === '' ? '*CURLIB' : lib });
+    pgm.addParam({ type: '5p0', value: data.length });
+    pgm.addParam({ type: `${data.length}A`, value: data });
+
+    this.conn.add(pgm.toXML());
+    let rtValue; // The returned value.
+    const toJson = (transportError, str) => {
+      if (transportError) {
+        if (this.reportError) {
+          cb(transportError, null);
+          return;
+        }
+        cb(null);
+        return;
+      }
+      parseString(str, (parseError, result) => {
+        if (parseError) {
+          if (this.reportError) {
+            cb(parseError, null);
+            return;
+          }
+          cb(str);
+          return;
+        }
+        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
+          rtValue = true;
+        } else {
+          rtValue = str;
+        }
+        if (this.reportError) {
+          cb(null, rtValue); // Run the call back function against the returned value.
+          return;
+        }
+        cb(rtValue);
+      });
+    };
+
+    this.conn.run(toJson); // Post the input XML and get the response.
   }
 
+  /**
+   * @param {string} name
+   * @param {string} lib
+   * @param {string} length
+   * @param {function} cb
+   */
   receiveFromDataQueue(name, lib, length, cb) {
-    iDataQueueDeprecate('As of v1.0, \'iDataQueue.receiveFromDataQueue()\' is deprecated. Please use \'Toolkit.receiveFromDataQueue()\' instead.');
-    this.toolkit.receiveFromDataQueue(name, lib, length, cb);
+    iDataQueueDeprecate('As of v1.0, \'iDataQueue.receiveFromDataQueue()\' is deprecated and will be removed at a later time.');
+    const pgm = new ProgramCall('QRCVDTAQ', { lib: 'QSYS' });
+    pgm.addParam({ type: '10A', value: name });
+    pgm.addParam({ type: '10A', value: lib === '' ? '*CURLIB' : lib });
+    pgm.addParam({ type: '5p0', value: length });
+    pgm.addParam({ type: `${length + 1}A`, value: '' });
+    pgm.addParam({ type: '5p0', value: 0 });
+
+    this.conn.add(pgm.toXML());
+    let rtValue; // The returned value.
+
+    const toJson = (transportError, str) => {
+      if (transportError) {
+        if (this.reportError) {
+          cb(transportError, null);
+          return;
+        }
+        cb(null);
+        return;
+      }
+      parseString(str, (parseError, result) => {
+        if (parseError) {
+          if (this.reportError) {
+            cb(parseError, null);
+            return;
+          }
+          cb(str);
+          return;
+        }
+        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
+          rtValue = result.myscript.pgm[0].parm[3].data[0]._;
+        } else {
+          rtValue = str;
+        }
+
+        if (this.reportError) {
+          cb(null, rtValue);
+          return;
+        }
+        cb(rtValue);
+      });
+    };
+    this.conn.run(toJson); // Post the input XML and get the response.
   }
 
+  /**
+   * @param {string} name
+   * @param {string} lib
+   * @param {function} cb
+   */
   clearDataQueue(name, lib, cb) {
-    iDataQueueDeprecate('As of v1.0, \'iDataQueue.clearDataQueue()\' is deprecated. Please use \'Toolkit.clearDataQueue()\' instead.');
-    this.toolkit.clearDataQueue(name, lib, cb);
+    iDataQueueDeprecate('As of v1.0, \'iDataQueue.clearDataQueue()\' is deprecated and will be removed at a later time.');
+    const pgm = new ProgramCall('QCLRDTAQ', { lib: 'QSYS' });
+    pgm.addParam({ type: '10A', value: name });
+    pgm.addParam({ type: '10A', value: lib === '' ? '*CURLIB' : lib });
+
+    this.conn.add(pgm.toXML());
+    let rtValue; // The returned value.
+
+    const toJson = (transportError, str) => {
+      if (transportError) {
+        if (this.reportError) {
+          cb(transportError, null);
+          return;
+        }
+        cb(null);
+        return;
+      }
+      parseString(str, (parseError, result) => {
+        if (parseError) {
+          if (this.reportError) {
+            cb(parseError, null);
+            return;
+          }
+          cb(str);
+          return;
+        }
+        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
+          rtValue = true;
+        } else {
+          rtValue = str;
+        }
+        if (this.reportError) {
+          cb(null, rtValue);
+          return;
+        }
+        cb(rtValue);
+      });
+    };
+
+    this.conn.run(toJson); // Post the input XML and get the response.
   }
 }
