+++ lib/Toolkit.js
@@ -104,136 +104,11 @@ class Toolkit {
     };
   }
 
-  sendToDataQueue(name, lib, data, cb) {
-    const pgm = new ProgramCall('QSNDDTAQ', { lib: 'QSYS' });
-    pgm.addParam({ type: '10A', value: name });
-    pgm.addParam({ type: '10A', value: lib === '' ? '*CURLIB' : lib });
-    pgm.addParam({ type: '5p0', value: data.length });
-    pgm.addParam({ type: `${data.length}A`, value: data });
 
-    this.conn.add(pgm.toXML());
-    let rtValue; // The returned value.
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes(`+++ success`)) {
-          rtValue = true;
-        } else {
-          rtValue = str;
-        }
-        if (this.reportError) {
-          cb(null, rtValue); // Run the call back function against the returned value.
-          return;
-        }
-        cb(rtValue);
-      });
-    };
 
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
 
-  receiveFromDataQueue(name, lib, length, cb) {
-    const pgm = new ProgramCall('QRCVDTAQ', { lib: 'QSYS' });
-    pgm.addParam({ type: '10A', value: name });
-    pgm.addParam({ type: '10A', value: lib === '' ? '*CURLIB' : lib });
-    pgm.addParam({ type: '5p0', value: length });
-    pgm.addParam({ type: `${length + 1}A`, value: '' });
-    pgm.addParam({ type: '5p0', value: 0 });
 
-    this.conn.add(pgm.toXML());
-    let rtValue; // The returned value.
 
-    const toJson = (transportError, str) => {
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          rtValue = result.myscript.pgm[0].parm[3].data[0]._;
-        } else {
-          rtValue = str;
-        }
-
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
-
-  clearDataQueue(name, lib, cb) {
-    const pgm = new ProgramCall('QCLRDTAQ', { lib: 'QSYS' });
-    pgm.addParam({ type: '10A', value: name });
-    pgm.addParam({ type: '10A', value: lib === '' ? '*CURLIB' : lib });
-
-    this.conn.add(pgm.toXML());
-    let rtValue; // The returned value.
-
-    const toJson = (transportError, str) => {
-      // Convert the XML output into JSON
-      if (transportError) {
-        if (this.reportError) {
-          cb(transportError, null);
-          return;
-        }
-        cb(null);
-        return;
-      }
-      parseString(str, (parseError, result) => {
-        if (parseError) {
-          if (this.reportError) {
-            cb(parseError, null);
-            return;
-          }
-          cb(str);
-          return;
-        }
-        if (result.myscript.pgm[0].success && result.myscript.pgm[0].success[0].includes('+++ success')) {
-          rtValue = true;
-        } else {
-          rtValue = str;
-        }
-        if (this.reportError) {
-          cb(null, rtValue);
-          return;
-        }
-        cb(rtValue);
-      });
-    };
-
-    this.conn.run(toJson); // Post the input XML and get the response.
-  }
 
   getTCPIPAttr(cb) {
     const outBuf = {
